# Optimizations for GDScript VM