* Removed instruction argument count and instruction prefetching. This is now done on the fly. Reduces jumps.
* Removed the address mode switch, addressing is now simple array indexing.
* OPCODE_DISPATCH now goes directly to the next instruction, like in Godot 3.x.

Performance seems to improve significantly (more than 40% faster on release on typed code vs master), here is some quick benchmark I made:

**Master:**
Debug:
Time taken (untyped): 33.708msec.
Time taken (typed): 20.464msec.
Release:
Time taken (untyped): 8.257msec.
Time taken (typed): 5.761msec.

**This PR:**
Debug:
Time taken (untyped): 29.234msec.
Time taken (typed): 16.027msec.
Release:
Time taken (untyped): 6.41msec.
Time taken (typed): 3.311msec.

**Test source code:**

```GDScript
const elems = 65536

func benchmark():

	var arr = []
	for x in range(elems):
		arr.append(randi() % elems)


	var t = Time.get_ticks_usec()

	var acc = 0.0

	for e in arr:
		var e2 = arr[e]
		var e3 = arr[e2]
		acc += e * e2
		acc *= e3 + e
		acc = sqrt(acc)

	
	print("Time taken (untyped): ",(Time.get_ticks_usec() - t)/1000.0,"msec.")		

	var arr2 := PackedInt32Array()
	for x in range(elems):
		arr2.append(randi() % elems)


	t = Time.get_ticks_usec()

	var acc2 : float = 0.0

	for e in arr2:
		var e2 :int = arr2[e]
		var e3 :int = arr2[e2]
		acc2 += e * e2
		acc2 *= e3 + e
		acc2 = sqrt(acc)

	
	print("Time taken (typed): ",(Time.get_ticks_usec() - t)/1000.0,"msec.")		
```
I don't have much else I can use to test performance, so if anyone wants to lend a hand and compare with master (both on debug and release) or test in your production project, it would be very welcome.