@Calinou 2023-01-02
If anyone's curious, I ported the benchmark code to PHP and Python. Benchmark times were measured on an i7-6700K @ 4.4 GHz on Fedora 36.

### PHP

Time taken (PHP 8.1.13 + XDebug): 16.055 msec
Time taken (PHP 8.1.13): 3.538 msec
Time taken (PHP 8.1.13 + [JIT](https://stitcher.io/blog/php-8-jit-setup)): 1.605 msec

<details>
<summary>Code</summary>

```php
<?php

const elems = 65536;

function benchmark() {
	$arr = [];
	for ($i = 0; $i < elems; $i++) {
		$arr[] = mt_rand(0, 2147483647) % elems;
    }

	$t = microtime(true);

	$acc = 0.0;

	foreach ($arr as $e) {
		$e2 = $arr[$e];
		$e3 = $arr[$e2];
		$acc += $e * $e2;
		$acc *= $e3 + $e;
		$acc = sqrt($acc);
    }

	echo "Time taken (PHP): " . (microtime(true) - $t) * 1000 . " msec.";
}

benchmark();
```
</details>

### Python

Time taken (Python 3.9.0): 12.492 msec
Time taken ([Pyston](https://github.com/pyston/pyston) 2.3.4): 8.459 msec

<details>
<summary>Code</summary>

```python
import math
import random
from time import time

elems = 65536

def benchmark():
	arr = []
	for x in range(elems):
		arr.append(random.randint(0, 2147483647) % elems)

	t = time()

	acc = 0.0

	for e in arr:
		e2 = arr[e]
		e3 = arr[e2]
		acc += e * e2
		acc *= e3 + e
		acc = math.sqrt(acc)


	print("Time taken (Python):", (time() - t) * 1000, "msec.")

benchmark()
```
</details>