@Byteron 2023-01-02
> If anyone's curious, I ported the benchmark code to PHP, Python and Ruby. Benchmark times were measured on an i7-6700K @ 4.4 GHz on Fedora 36.
> 
> ## PHP
> * Time taken (PHP 8.1.13 + XDebug): 16.055 msec
> * Time taken (PHP 8.1.13): 3.538 msec
> * Time taken (PHP 8.1.13 + [JIT](https://stitcher.io/blog/php-8-jit-setup)): 1.605 msec[1](#user-content-fn-1-b653e5b2f1b72dcbf21f8402853ecf95)
> 
> Code
> ## Python
> * Time taken (Python 3.9.0): 12.492 msec
> * ~Time taken ([Pyston](https://github.com/pyston/pyston) 2.3.4): 8.459 msec[1](#user-content-fn-1-b653e5b2f1b72dcbf21f8402853ecf95)~
> * Time taken ([Pyston](https://github.com/pyston/pyston) 2.3.5): 6.686 msec[1](#user-content-fn-1-b653e5b2f1b72dcbf21f8402853ecf95)
> 
> Code
> ## Ruby
> * Time taken (Ruby 3.2.0): 7.357 msec
> * Time taken (Ruby 3.2.0 + `--jit`): 4.711 msec[1](#user-content-fn-1-b653e5b2f1b72dcbf21f8402853ecf95)
> 
> Code
> ## Footnotes
> 1. A JIT compiler is not comparable to interpreters in terms of performance. This figure is included for reference only. [↩](#user-content-fnref-1-b653e5b2f1b72dcbf21f8402853ecf95) [↩2](#user-content-fnref-1-2-b653e5b2f1b72dcbf21f8402853ecf95) [↩3](#user-content-fnref-1-3-b653e5b2f1b72dcbf21f8402853ecf95) [↩4](#user-content-fnref-1-4-b653e5b2f1b72dcbf21f8402853ecf95)

Given you are on different Hardware it probably makes sense to also include GDScript Numbers on your 6700K.