@bruvzg 2023-01-02
Some tests on M1 mac, using release editor binary (average of 10 runs):

|                          | Time    |
|--------------------------|---------|
| Godot (master, untyped)  | 13.9504 |
| Godot (master, typed)    |  8.9020 |
| Godot (this pr, untyped) | 11.2082 |
| Godot (this pr, typed)   |  6.0292 |
| Python 3.10.9            |  9.1007 |
| Ruby 2.6.10p210          |  8.6488 |