@Calinou 2023-01-02
> I'm new to godot, so can someone pls explain why the comparisons are made against python, php and ruby instead of c# and c++ since these are the languages someone might actually use in a godot project?

C# and C++ are compiled languages, so these aren't fair comparisons to make. It's not possible for an interpreted language to even match a compiled language.