@jordo 2023-01-02
> I'm new to godot, so can someone pls explain why the comparisons are made against python, php and ruby instead of c# and c++ since these are the languages someone might actually use in a godot project?

We've been doing a lot of transpiling gdscript to c++ for our game logic in some performance heavy areas... so I have a lot of experience here.  We run a lot of godot gameservers on cloud infrastruction so we pay for every CPU cycle used.  

For GDScript vs C++ we can typically see close to two orders of magnitude of speedup, (usually 40-100X faster), which is why it's not a fair comparison but also illustrates how there can be so much to gain with an AOT language + optimization pass that compiles to machine instructions.

The typed gdscript example code runs in `16.033msec` on my 2019 x86 macbook pro.   

The native c++ version below, executes in `0.021000msec`.  So in this example, actually close to 1000X faster.

```
	int elems = 65536;
	
	std::vector<int> arr = {};
	for (int i = 0; i < elems; i++) {
		arr.push_back(rand() % elems);
	}

	uint64_t t = mach_absolute_time();

	float acc = 0.0;

	for (int e = 0; e < elems; e++) {
		int e2 = arr[e];
		int e3 = arr[e2];
		acc += e * e2;
		acc *= e3 + e;
		acc = sqrt(acc);
	}

	printf("Time taken (typed): %.4f%s",(mach_absolute_time() - t)/1000.0,"msec.");
```

Now we typically don't see a 1000X improvement.  But in this trivial example the native optimization passes that a c++ compiler can do (vectorization, inlining, loop unrolling, et al) seem to really play a factor.  In production we usually see about a 40X-100X performance improvement.