@reduz 2023-01-02
@jordo We have been discussing this here - https://github.com/godotengine/godot-proposals/issues/6031 and the general consensus seems to be that users would be happy with an optional transpiler to C, but this is obviously Godot 4.0+ material.